requirements.txt
google-cloud-aiplatform
google-adk
fastapi
uvicorn
chromadb
pypdf
python-dotenv
tiktoken
pytest